Fix #54: use `jupyter labhub` to start jupyterlab
This is the proper incantation for using jupyter server, lab, and hub

labhub mixes in a jupyter_server.serverapp.ServerApp
https://github.com/jupyterlab/jupyterlab/blob/239398c7ffdc37a169d584b5fb3ea9068a71c983/jupyterlab/labhubapp.py
with a singleuser_app.

Just calling jupyterhub-singleuser directly uses make_singleuser_app
which has no knowledge of the new jupyter_server so the configuration
specified in jupyter_server_config.py (allowed_kernelspecs) was not
being picked up.

Also fix #29. jupyter-server was released in #53 but this actually
uses it.